Move typeform, s3, fb functions into own files
