+++ package.json
@@ -14,33 +14,33 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.2.2",
-    "@babel/node": "^7.0.0",
-    "@babel/preset-env": "^7.2.3",
-    "babel-loader": "^8.0.5",
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/node": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
+    "babel-loader": "^8.0.6",
     "node-schedule": "^1.3.2",
-    "webpack": "^4.28.4",
-    "webpack-cli": "^3.2.1",
-    "webpack-dev-middleware": "^3.5.0",
-    "webpack-hot-middleware": "^2.24.3",
+    "webpack": "^4.34.0",
+    "webpack-cli": "^3.3.4",
+    "webpack-dev-middleware": "^3.7.0",
+    "webpack-hot-middleware": "^2.25.0",
     "ws": "3.3.2"
   },
   "dependencies": {
-    "@sendgrid/mail": "^6.3.1",
-    "aws-sdk": "^2.466.0",
+    "@sendgrid/mail": "^6.4.0",
+    "aws-sdk": "^2.476.0",
     "babel-polyfill": "^6.26.0",
     "bcryptjs": "^2.4.3",
-    "bluebird": "^3.5.3",
-    "body-parser": "^1.18.3",
+    "bluebird": "^3.5.5",
+    "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "cron": "^1.7.1",
     "dotenv": "^7.0.0",
-    "express": "^4.16.4",
+    "express": "^4.17.1",
     "fb-messenger": "^2.3.1",
-    "jsonwebtoken": "^8.4.0",
+    "jsonwebtoken": "^8.5.1",
     "mime-types": "^2.1.24",
-    "mysql2": "^1.6.4",
+    "mysql2": "^1.6.5",
     "node-schedule": "^1.3.2",
     "paypal-rest-sdk": "^1.8.1",
     "request": "^2.88.0",
