+++ src/server/routes/refugeeProtected.js
@@ -1,15 +1,16 @@
 import express from "express";
-import controller from "./../controller/refugee";
+// import controller from "./../controller/refugee";
+import typeformController from "./../controller/typeform";
 
 const router = express.Router();
 const app = express();
 
 router.post("/typeformV3", (req, res) => {
-  controller.processTypeformV3(req, res);
+  typeformController.processTypeformV3(req, res);
 });
 
 router.post("/typeformV4", (req, res) => {
-  controller.processTypeformV4(req, res);
+  typeformController.processTypeformV4(req, res);
 });
 
 export default router;
